Reorganise Provider Interface
